requirements.txt
qrcode~=7.4.2
python-dotenv~=1.0.0
fastapi~=0.94.1
pydantic~=1.10.6